Move GraphExplorer up
